Move ApiErrorboundary to components/common and cleanup
