--- airbyte-webapp/src/components/ApiErrorBoundary/index.tsx
@@ -1,3 +1,0 @@
-import ApiErrorBoundary from "./ApiErrorBoundary";
-
-export default ApiErrorBoundary;
